// docs/site/src/paths.gen.ts
/**
 * GENERATED by last-ts file router — do not edit.
 */
export const fileEntries = [
  {
    "id": "index",
    "filePath": "/",
    "routePath": "/"
  },
  {
    "id": "404",
    "filePath": "/404",
    "routePath": "/404"
  },
  {
    "id": "api",
    "filePath": "/api",
    "routePath": "/api"
  },
  {
    "id": "api_pkg",
    "filePath": "/api/[pkg]",
    "routePath": "/api/:pkg"
  },
  {
    "id": "api_pkg_module",
    "filePath": "/api/[pkg]/[module]",
    "routePath": "/api/:pkg/:module"
  },
  {
    "id": "api_pkg_module_symbol",
    "filePath": "/api/[pkg]/[module]/[symbol]",
    "routePath": "/api/:pkg/:module/:symbol"
  },
  {
    "id": "api_hyperlink-ts_module_symbol",
    "filePath": "/api/hyperlink-ts/[module]/[symbol]",
    "routePath": "/api/hyperlink-ts/:module/:symbol"
  },
  {
    "id": "docs_chapter",
    "filePath": "/docs/[chapter]",
    "routePath": "/docs/:chapter"
  },
  {
    "id": "docs_hyperlinks",
    "filePath": "/docs/hyperlinks",
    "routePath": "/docs/hyperlinks"
  },
  {
    "id": "docs_resources",
    "filePath": "/docs/resources",
    "routePath": "/docs/resources"
  },
  {
    "id": "releases",
    "filePath": "/releases",
    "routePath": "/releases"
  },
  {
    "id": "search",
    "filePath": "/search",
    "routePath": "/search"
  }
] as const;

export const filePaths = [
  "/",
  "/404",
  "/api",
  "/api/[pkg]",
  "/api/[pkg]/[module]",
  "/api/[pkg]/[module]/[symbol]",
  "/api/hyperlink-ts/[module]/[symbol]",
  "/docs/[chapter]",
  "/docs/hyperlinks",
  "/docs/resources",
  "/releases",
  "/search",
] as const;

export const routePaths = [
  "/",
  "/404",
  "/api",
  "/api/:pkg",
  "/api/:pkg/:module",
  "/api/:pkg/:module/:symbol",
  "/api/hyperlink-ts/:module/:symbol",
  "/docs/:chapter",
  "/docs/hyperlinks",
  "/docs/resources",
  "/releases",
  "/search",
] as const;

export type FilePath = "/" | "/404" | "/api" | "/api/[pkg]" | "/api/[pkg]/[module]" | "/api/[pkg]/[module]/[symbol]" | "/api/hyperlink-ts/[module]/[symbol]" | "/docs/[chapter]" | "/docs/hyperlinks" | "/docs/resources" | "/releases" | "/search";
export type RoutePath = "/" | "/404" | "/api" | "/api/:pkg" | "/api/:pkg/:module" | "/api/:pkg/:module/:symbol" | "/api/hyperlink-ts/:module/:symbol" | "/docs/:chapter" | "/docs/hyperlinks" | "/docs/resources" | "/releases" | "/search";
export type FileRouteId = "index" | "404" | "api" | "api_pkg" | "api_pkg_module" | "api_pkg_module_symbol" | "api_hyperlink-ts_module_symbol" | "docs_chapter" | "docs_hyperlinks" | "docs_resources" | "releases" | "search";
export type FileEntry = (typeof fileEntries)[number];